requirements.txt
dj-database-url==0.4.1
Django==1.9.6
django-enumfields==0.7.4
django-floppyforms==1.6.2
psycopg2==2.6.1
six==1.10.0
whitenoise==3.0
